# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Voice Receptionist API
  version: 0.1.0
  description: >
    Covers the surface built in this session (Modules 1-4: platform
    foundation, authentication/authorization, and the scheduling domain).
    Call/knowledge/escalation/usage/retention endpoints (Modules 5-11) do
    not exist yet — see docs/adr/dialect-feasibility-verdict.md — and are
    not documented here.
servers:
  - url: http://localhost:3001
    description: Local development (apps/api)
security:
  - sessionCookie: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-session
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          example: VALIDATION_FAILED
        message:
          type: string
        correlationId:
          type: string
        details:
          type: object
...
paths:
...
  /tenants/{tenantId}:
    get:
      summary: Get tenant profile (A3.7 - Tenant profile R).
      parameters: [{ $ref: "#/components/parameters/tenantId" }]
      responses:
        "200": { description: OK }
        "403": { $ref: "#/components/responses/Forbidden" }
    patch:
      summary: Update tenant profile (tenant_owner only).
      parameters: [{ $ref: "#/components/parameters/tenantId" }]
      responses:
        "200": { description: OK }
        "403": { $ref: "#/components/responses/Forbidden" }
  /tenants/{tenantId}/users:
    get:
      summary: List tenant users (tenant_owner full; platform_operator identity-only).
      parameters: [{ $ref: "#/components/parameters/tenantId" }]
      responses:
        "200": { description: OK }
    post:
      summary: Create a tenant user (tenant_owner only).
      parameters: [{ $ref: "#/components/parameters/tenantId" }]
      responses:
        "201": { description: Created }
...
  /tenants/{tenantId}/appointments:
    get:
      summary: List appointments for a location.
      responses:
        "200": { description: OK }
    post:
      summary: >
        Book an appointment (A7.1: a voice_call source must carry
        callerConfirmed: true). Concurrency-safe via the database
        exclusion constraint (A4.3) — a contended slot returns
        SLOT_CONTENTION, never a silent double-booking.
      responses:
        "201": { description: Created }
        "409":
          description: SLOT_CONTENTION — the slot was booked by a concurrent request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /tenants/{tenantId}/appointments/{appointmentId}/reschedule:
    patch:
      summary: Reschedule an appointment; may return SLOT_CONTENTION (A7.3).
      responses:
        "200": { description: OK }
        "409": { description: SLOT_CONTENTION }
  /tenants/{tenantId}/appointments/{appointmentId}/status:
    patch:
      summary: >
        Transition an appointment's status. Every transition writes an
        audit log entry naming the actor, principal type, and source
        (A4.4).
      responses:
        "200": { description: OK }
        "409": { description: Invalid status transition }
...
components:
  parameters:
    tenantId:
      name: tenantId
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Forbidden:
      description: A3.6 - forbidden operation, no resource data included in the body
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }